# Dynamic kernel selection for traning/inference