Right now, `aqlm` requires one to load models with
```python
with aqlm.optimize_for_training():
    model = AutoModelForCausalLM.from_pretrained(...)
```
to use kernels optimized for larger batch sizes. It doesn't need to be that way since it's pretty easy to tell which kernels should be used by just looking at the input.

This PR does that.