@BlackSamorez 2024-03-05
This is what throughput looked like before this PR for a `4096x12288` matrix on an _RTX 3090_. Based on that, I put the matvec/matmat threshold at 32 tokens.
| num_tokens (batch_size x seq_len) | with `optimize_for_training`, ms/pass | without `optimize_for_training`, ms/pass | `fp16` baseline |
|--|--|--|--|
| 1 | 0.21 | 0.18 | 0.14 |
| 4 | 0.54 | 0.54 | 0.14 |
| 16 | 1.92 | 1.92 | 0.14 |
| 64 | 4.74 | 4.73 | 0.16 |
| 256 | 5.03 | too slow | 0.45 |
| 1024 | 6.13 | too slow | 1.57 |
| 4096 | 10.08 | too slow | 5.54 |
| 16384 | 25.70 | too slow | 21.14 |